% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arxiv_batches.R
\name{arxiv_batches}
\alias{arxiv_batches}
\title{Query the arXiv API in batches.}
\usage{
arxiv_batches(query, start, limit, batch_size, sleep, timeout = 30, sep = "|")
}
\arguments{
\item{query}{String with the API query}

\item{start}{Index to start at.}

\item{limit}{Maximum number of papers to return.}

\item{batch_size}{batch_size}

\item{sleep}{Seconds to sleep between batches}

\item{timeout}{Timeout seconds}

\item{sep}{Separator}
}
\value{
A tibble of papers
}
\description{
Will continue sleeping, then attempting to retrieve the batch until a result
the same size as the batch is found.
}
\examples{
\donttest{
arxiv_batches(
  query = "cat:cs.AI",
  start = 0,
  limit = 5,
  batch_size = 1,
  sleep = 1
)
}
}
