% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arxiv_request.R
\name{arxiv_request}
\alias{arxiv_request}
\title{Query the arXiv API.}
\usage{
arxiv_request(
  query,
  start = 0,
  limit,
  batch_size,
  sleep = 0,
  timeout = 30,
  sep = "|"
)
}
\arguments{
\item{query}{String with the API query}

\item{start}{Index to start at}

\item{limit}{Maximum number of papers to return}

\item{batch_size}{batch_size}

\item{sleep}{Seconds to sleep between batches}

\item{timeout}{Timeout seconds}

\item{sep}{Separator}
}
\value{
A tibble
}
\description{
For instructions on writing queries, see the aRxiv package tutorial:
https://ropensci.org/tutorials/arxiv_tutorial/, specifically the "Search terms"
section.
}
\examples{
\donttest{
arxiv_request(
  query = "cat:cs.AI",
  start = 0,
  limit = 5,
  batch_size = 1,
  sleep = 1
)
}
}
