% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arxiv_batches.R
\name{arxiv_sleep_then_batch}
\alias{arxiv_sleep_then_batch}
\title{Sleep, then retrieve batch}
\usage{
arxiv_sleep_then_batch(
  query,
  start,
  limit,
  batch_size,
  batch_number,
  sleep,
  timeout = 30,
  sep = "|"
)
}
\arguments{
\item{query}{String with the API query}

\item{start}{Index to start at.}

\item{limit}{Maximum number of papers to return.}

\item{batch_size}{Batch size}

\item{batch_number}{Batch index}

\item{sleep}{Seconds to sleep between batches}

\item{timeout}{Timeout seconds}

\item{sep}{Separator}
}
\value{
A tibble
}
\description{
Sleep, then retrieve batch
}
\examples{
\donttest{
arxiv_sleep_then_batch(
  query = "cat:cs.AI",
  start = 0,
  limit = 5,
  batch_size = 1,
  sleep = 1
)
}
}
